# ci: Speed up playwright tests