--- a/.github/workflows/validate.yml
+++ b/.github/workflows/validate.yml
@@ -70,7 +70,7 @@ jobs:
         run: pnpm test:unit
   playwright:
     name: Playwright tests
-    runs-on: macos-11
+    runs-on: macos-13
     env:
       CI: true
     steps: