Turns out [we've been only using 1 worker for playwright tests on CI](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210325096/job/22457550534#step:8:13). As a result, our playwright test job tends to take anywhere from 6-10 minutes. It's by far the bottleneck in CI: all our other workflows take <2 minutes.

The first change I made was to explicitly set `maxWorkers` in the playwright config so we can utilize all cores on CI. I tested using all cores locally on my MBP (10 cores) and Linux desktop (8 cores), but neither seemed very stable, so I've set `maxWorkers` to `undefined` locally, resulting in the default 4 cores.

Additionally, testing out newer `macos` runners, it seems like the latest `macos-14` runners are the fastest, with job times averaging 2.5 minutes, likely because `macos-14` runners use M1 rather than Intel.

`macos-14` runners _are_ in beta, but they seem stable enough. If we're not comfortable using a beta runner, `macos-13` or `macos-latest` runners are still decent when utilizing all cores. Using `macos-latest` _would_ be the least maintenance burden as it's a moving target (currently an alias for `macos-12`).

[Info about github-hosted runners](https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories).

`macos-14` run times (**avg. 2m 28s**):
- [2m 48s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210323909/attempts/1)
- [2m 21s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210323909/attempts/2)
- [2m 16s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210323909/attempts/3)
- [2m 22s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210323909/attempts/4)
- [3m 21s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210323909/attempts/5)

`macos-13` run times (**avg. 4m 34s**):
- [4m 34s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8211982450/attempts/1)
- [4m 32s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8211982450/attempts/2)
- [4m 59s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8211982450/attempts/3)
- [4m 37s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8211982450/attempts/4)
- [4m 10s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8211982450/attempts/5)

`macos-latest` (currently `macos-12`) run times (**avg. 5m 1s**):
- [6m 28s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210346905/attempts/1)
- [4m 13s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210346905/attempts/2)
- [6m 9s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210346905/attempts/3)
- [4m 13s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210346905/attempts/4)
- [4m 6s](https://github.com/vanilla-extract-css/vanilla-extract/actions/runs/8210346905/attempts/5)